Adds FlakyActivityTestRule with @Repeat and @AllowFlaky annotations (#23)
* Adds FlakyActivityTestRule and @Repeat and @AllowFlaky annotations

This subclass of the ActivityTestRule let's you add two different annotations
to your test methods to execute them several times and either allow some flakyness
or avoid it.

* Fixed some pmd and checkstyle violations

* Merged AllowFlakyRule and RepeatRule into FlackyActivityTestRule

We didn't need them separated, and combining them allows us to make some extra validations.
I also added a few unit tests for the creation logic of the statement.

NOTE: This change removes, for now, the behaviour of AllowFlaky enabled by default without the annotation. It will be introduced better later.

* Extracted statement creation logic into a Builder

This FlakyActivityStatementBuilder is an abstraction of the FlakyActivityTestRule for easier testing and to allow more features soon.

* Adds possibility to use AllowFlaky by default on all test methods

There is a convenience method in the FlakyActivityTestRule to enable default flaky attempts
in case a test method doesn't have neither the @Repeat or the @AllowFlaky annotations.

This to avoid repeating the annotation in a big codebase with lots of flaky tests.
Espresso is flaky. Deal with it!

* Added javadoc

* Added javadoc for allowFlakyAttemptsByDefault

* Added usage of FlakyActivityTestRule to the Readme

* Replaced TenRepetitionsActivityTestRule with FlakyActivityTestRule

* Renamed FlakyActivity to HelloWorldActivity

* Reformat to comply with checkstyle