Create BaristaRule to rule them all
This rule includes all other rules, so it's easier and simpler to write tests without boilerplate.
I wasn't sure how to add tests for it... Repeating the same tests that all rules have it's a lot of duplication. And I don't know how to test that a rule includes other rules.

I just tested manually by changing the current rules tests and replacing the original rules with this one.
It works.