Remove api limit to PermissionGranter
The PermissionGranter class is annotated with RequiresApi 18 (because UI automator requires that api level).
This means that using PermissionGranter in an application that supports APIs lower than 18, Lint will give an error.

But in sdks lower than 23 we don't even use UiAutomator because permissions are granted at install time, so we're actually fine.

This PR removes the annotation so that we don't misslead the user.