Migrate ClearDatabaseRule to Kotlin
I refactored the responsabilities of the rule and the DatabaseOperations file. The latter is now a simple collaborator to abstract from Android's SQLite database, instead of doing all the work.
The rule logic is now on the rule itself.

I added unit tests for the behavior of the rule. It wasn't really possible before because DatabaseOperations was made of static methods.