Exclude kotlin files from Findbugs (#126)
This needs to happen.
Kotlin generates code that doesn't make findbugs happy, but we don't care about his feelings,
we like Kotlin better. He's the new kid in the hood and we're friends.

No, really. We've done 2 PRs with kotlin code and findbugs failed in both because of the bytecode generation.
It's not helpful for kotlin files.