Improve error handling on magic methods (#134)
* Fix magic error handling

* Fix bug with click on half displayed views

When a view is half displayed and we try to click, the type of exception we get is different. It's not the matcher that fails, it's the action.
I've reproduced the issue with a half displayed button (by scrolling to half it's position) and fixed the code.

* Update EditText tests

Avoid using the flaky rule (Barista should not be flaky!) and add the FailureHandlerValidator test

* Increase scroll retries to 50

With fewer retries than 50 some tests fail 5% of times.
I reproduced this with EditText and Autocomplete tests.

It's not an ideal solution, but it works. I'm opening another issue in Barista to improve the scrollTo action.

* Composition for displayed matchers.

* Simplifly internals of ListViewActions

We were implementing the `clickListItemForMultipleListsOnScreen` wrong. It's for a very specific use-case in one of our internal apps, and it doesn't really do what it says.
Please lord forgive us, we were new to Espresso.

* Retry scroll 50 times instead of 51

* Error message with capital letter