Merge RecyclerView and ListView actions into ListActions (#146)
* Create BaristaListActions to perform actions on ListView and RecyclerView

Unlike BaristaListViewActions and BaristaRecylerViewActions, this new class will handle both list components without distinction to the user.
Also id is now optional. If there are more than one listview and/or recyclerview displayed it will be mandatory.

I created a ListsActivity to create layouts with different lists combinations dynamically from the test. This allows us to test multiple scenarios without having a different activity for each one.

* Add scroll action to BaristaListActions

* Extract common list action logic to performMagicAction function

The logic is always the same: verify a single match, try performing on a recycler, try performing on a listview
Also, the recycler and listview actions logic is the same respectively, so I also extracted them to common functions.

* Add actions to click on list childs

* Replace methods on RecyclerView with ViewPager test

* Delete old ListView and RecyclerView actions and tests

* Checkstyle violation

* Fix parameter ordering

* Change package visibility for findbugs

* Fix lint violation

* Fix PMD violations. I hate my life