Extract ugly try/catch code to a magicAssertOnView function and use it
I also created some infix extension functions to make it even easier to write. It extends any Matcher<View>, String or Int (resource).
I added a `not` operator to Matcher<View>.
I documented the behaviour the best I could.

I replaced the in-site code with this function, and applied it to other methods of the BaristaVisibilityAssertions that were not doing any magic.