Fix error handling when performing list action
We were missing a couple of tests to check what happens when the recycler or listview is found but the action performed fails.
Becase of this we had a bug, we were not notifying espresso the EspressoFailure, which breaks any custom FailureHandler setted by the user.
I added the tests and fixed the code.