Clear files with optional regex filter
Simply deleting all files might break things for some apps.
For example, a library might depend on some files that get stored when initialized only.
Or you might have a set of files that don't want deleted between test.

For that reason I added a regex filter to delete only a set of files.
I tried to mimic the api and implementation of ClearDatabaseRule, and added unit tests.