Update Readme about the new dependency coordinates
While preparing for the Maven Central migration I replaced the package name and dependency coordinates of the library. But this led to broken download instructions in the Readme! (See https://github.com/AdevintaSpain/Barista/issues/425) 

I'm adding some clarification about the different versions of Barista published in different places with different names. 
We should update the Maven Central instructions once the migration is complete. In the meantime, we should allow people to use the old dependency.